requirements.txt
requests
yfinance
pandas
google-generativeai
feedparser